Create guess_the_number
To play the game, run this script, and you'll be prompted to guess the randomly chosen number within a specified range. You can modify the lower_limit and upper_limit variables to change the range of numbers.

Have fun playing the game, and feel free to modify and expand it as desired!